basic adi methods, pulled from original code
This pulls the methods that used to be in a single file out into a more
namespaced format, with dataset, organization, and transformation as
top-level namespaces.

I've played with some name changes as I went along to better capture
what's going on. I'm not too sure I'm happy with them just yet. I'm
also thinking that a little OO may be in order. I was avoiding it until
now in the name of not over-engineering something simple, but there's at
least enough passing of the api connection information to justify it,
and it would also make it easier to hide some of the functionality
that's not intended to be exposed.

As for names, I'm trying 'define' for code-based creation... like
creating a reusable transformation: transformation.define... or a
code-based dataset: dataset.define, whereas raw 'create' is used for
basic dataset creation, and 'upload' is left for when you want to
create and upload to a dataset.