add python-api methods to use new server endpoints
This allows listing, getting properties of, and deleting top level
transformations. There's some duplication creeping in here that
I'm planning to resolve by introducing an intermediate API class.

That's still going to leave some duplication, though. This other
duplication is resulting from trying to deal with name uniqueness
constraints in two places. On the Python API side, we're doing
extra queries and contortions to prevent triggering GraphQL errors,
and on the GraphQL side, we're basically retracing a bunch of the
same steps to make sure that the user isn't trying to spoof something.

If we switched on the API side to requiring an organization prefix to
anything using a name instead of a uuid ('some_org:some_dataset' or
'some_org:SomeTransformation'), we'd be able to relieve some of
the burden on the client side and reduce some API parameter bloat.

I want to think on this a bit, and I certainly don't want to muck up
this PR with it.