add query method to the connection class
We don't want to always have to do a one-to-one addition of python-api
methods to match new overall API functionality just to be able to test
it. But the python-adi package is useful for testing. This exposes the
raw gql_query method, wrapped in a simple query method that uses the
connection info already established.