basic rudimentary testing
This defines a custom TestTransformation class so that we don't have to
worry about getting data from swift just yet and can test just the main
logic. Right now, we're just writing results to the console. Yet to be
done: Use the same python test package that ADI integration tests use.

First we test that we can use a "transformation" just like a normal
function. Then we test that we can pass parameters into a .run function
attached to it, just as ADI would. Finally, we're testing that we can
figure out, from just a block of code, where our actual
"transformations" are programatically. Whatever bootstraps the running
of these transformations will need to know that, after all.