make default a special key for loader param
Thinking ahead, if we're going to have to subclass Transformation to
provide custom behaviour, it's nice to have a simple, elegant
constructor. The lack of symmetry between loader/writer was weird.
It's still a bit weird, but at least this way we're not always having to
deal with an extra parameter.