update to work with more verbose json from ADI
In addition to being more verbose, it should be much more flexible,
allowing us to simply code up new Transformation subclasses (or
change the existing Transformation class) with more advanced
loaders to deal smartly with different formats and/or storage types.
Most importantly, it also leaves it open to a transformation writer
to specify their own behaviour. Although we're keeping test_loader
and test_writer simple, we could import the default functions and
defer to them if we're not getting the "test-local" format.