make @transformation actually work with ADI
There was a lot of experimentation and juggling here, and it's hard to
make the changes look pretty. The current tests define some of the
input/output expectations when transform code is run from ADI (as
opposed to a simple function call while testing in a Jupyter notebook).

There's a lot of thought towards the future here. In the end, the
decorated transformation function gets a .run method attached,
which ADI expects to be able to pass a strongly structured parameter
map. A .metadata variable is attached as well, allowing the enclosing
transformation class to write to it pretty much any way it wants (in
the default Transformation, we just run the results through some
pandas functions to collect column info). Finally, .inputs is attached
to record the results of any mappings made while defining the
transformation (for non-reusable transformations).

More explicit documentation to follow in the manual.