handle writing streamed output
This turned out to be a little trickier than anticipated, due to the
introduction of asyncio to deal with other problems. Luckily the httpx
library handles writing data that's an async generator.

With all of this in place, we're basically letting the output writing
iteration pull the input and trigger the transform. We're also able to
write output data, take samples of it, and collect metadata from the
entire dataset while still only reading it once.